refactor(api): add specific answer model to certification/evaluation
