+++ api/src/certification/evaluation/domain/models/Answer.js
@@ -1,0 +1,15 @@
+import { AnswerStatus } from '../../../../shared/domain/models/AnswerStatus.js';
+
+class Answer {
+  constructor({ id, result, challengeId } = {}) {
+    this.id = id;
+    this.result = AnswerStatus.from(result);
+    this.challengeId = challengeId;
+  }
+
+  isOk() {
+    return this.result.isOK();
+  }
+}
+
+export { Answer };
